Adding zip creation
When someone hits the build endpoint, we create the projects through the
manifestjs library to a tmp folder in tmp/<id>/<short_name>.  Once that
completes, then we zip of the entire folder.

Next steps are to upload the zip file to Azure and then return the file
url back to the requester.