Huge Refactor
Moved a lot of the logic for the routes into a seperate controller.
This enabled a split in dependencies between the controller and the
services it consumed, making it easier to test.

Added tests to cover the build steps of the controller in part.  Still
needs more work.

Slit tests into unit and acceptance tests.  Moved the util helper to the
acceptance folder as that is what it was made for.